@@ -4,28 +4,28 @@
     "description": "Sass and JS for Admin Columns",
     "private": true,
     "devDependencies": {
-        "@babel/core": "7.10.2",
-        "@babel/preset-typescript": "^7.3.3",
+        "@babel/core": "7.22.20",
+        "@babel/preset-typescript": "7.22.15",
         "@types/jquery": "^3.5.5",
-        "@types/node": "^14.14.37",
-        "axios": "^0.21.0",
-        "babel-loader": "^8.0.5",
+        "@types/node": "^20.6.2",
+        "axios": "1.5.0",
+        "babel-loader": "9.1.3",
         "body-scroll-lock": "^4.0.0-beta.0",
         "jquery": "^3.5.1",
         "js-cookie": "^2.2.1",
-        "nanobus": "4.4.0",
+        "nanobus": "4.5.0",
         "nps": "^5.9.12",
         "nps-utils": "^1.7.0",
         "sass": "^1.32.12",
-        "svelte": "^3.42.6",
-        "svelte-check": "^2.2.4",
-        "svelte-loader": "^3.1.2",
-        "svelte-preprocess": "^4.6.5",
+        "svelte": "4.2.0",
+        "svelte-check": "3.5.1",
+        "svelte-loader": "3.1.9",
+        "svelte-preprocess": "5.0.4",
         "svg-symbols": "1.0.4",
         "svgstore-cli": "1.3.1",
         "tailwindcss": "^3.2.4",
-        "ts-loader": "^8.0.14",
-        "typescript": "^4.2",
+        "ts-loader": "^9.4.4",
+        "typescript": "^5.2.2",
         "webfont": "8.1.4",
         "webpack": "^5.19.0",
         "webpack-cli": "^4.5.0",
