@@ -4,18 +4,18 @@ namespace AC\Column\Comment;
 
 use AC\Column;
 
-/**
- * @since 2.0
- */
-class Date extends Column {
+class Date extends Column
+{
 
-	public function __construct() {
-		$this->set_original( true )
-		     ->set_type( 'date' );
-	}
+    public function __construct()
+    {
+        $this->set_original(true)
+             ->set_type('date');
+    }
 
-	public function register_settings() {
-		$this->get_setting( 'width' )->set_default( 14 );
-	}
+    public function register_settings()
+    {
+        $this->get_setting('width')->set_default(14);
+    }
 
 }
