@@ -5,28 +5,29 @@ namespace AC\Column\Comment;
 use AC;
 use AC\Column;
 
-/**
- * @since 2.4.7
- */
-class Post extends Column {
+class Post extends Column
+{
 
-	public function __construct() {
-		$this->set_type( 'column-post' );
-		$this->set_label( __( 'Post', 'codepress-admin-columns' ) );
-	}
+    public function __construct()
+    {
+        $this->set_type('column-post');
+        $this->set_label(__('Post', 'codepress-admin-columns'));
+    }
 
-	public function get_raw_value( $id ) {
-		$comment = get_comment( $id );
+    public function get_raw_value($id)
+    {
+        $comment = get_comment($id);
 
-		if ( ! $comment ) {
-			return false;
-		}
+        if ( ! $comment) {
+            return false;
+        }
 
-		return $comment->comment_post_ID;
-	}
+        return $comment->comment_post_ID;
+    }
 
-	public function register_settings() {
-		$this->add_setting( new AC\Settings\Column\Post( $this ) );
-	}
+    public function register_settings()
+    {
+        $this->add_setting(new AC\Settings\Column\Post($this));
+    }
 
 }
