@@ -2,53 +2,47 @@
 
 namespace AC\Helper\Select;
 
-final class OptionGroup {
+final class OptionGroup
+{
 
-	/**
-	 * @var string
-	 */
-	private $label;
+    private $label;
 
-	/**
-	 * @var Option[]
-	 */
-	private $options;
+    /**
+     * @var Option[]
+     */
+    private $options;
 
-	/**
-	 * @param string   $label
-	 * @param Option[] $options
-	 */
-	public function __construct( $label, array $options = [] ) {
-		$this->label = $label;
+    /**
+     * @param string   $label
+     * @param Option[] $options
+     */
+    public function __construct(string $label, array $options = [])
+    {
+        $this->label = $label;
 
-		foreach ( $options as $option ) {
-			$this->add_option( $option );
-		}
-	}
+        foreach ($options as $option) {
+            $this->add_option($option);
+        }
+    }
 
-	/**
-	 * @return string
-	 */
-	public function get_label() {
-		return $this->label;
-	}
+    public function get_label(): string
+    {
+        return $this->label;
+    }
 
-	/**
-	 * @return Option[]
-	 */
-	public function get_options() {
-		return $this->options;
-	}
+    /**
+     * @return Option[]
+     */
+    public function get_options(): array
+    {
+        return $this->options;
+    }
 
-	/**
-	 * @param Option $option
-	 *
-	 * @return $this
-	 */
-	protected function add_option( Option $option ) {
-		$this->options[] = $option;
+    protected function add_option(Option $option): self
+    {
+        $this->options[] = $option;
 
-		return $this;
-	}
+        return $this;
+    }
 
 }
