@@ -19,34 +19,9 @@ final class Response {
 		$this->more = $more;
 	}
 
-	private function parse_options( array $options ): array {
-		$results = [];
-
-		foreach ( $options as $option ) {
-			switch ( true ) {
-				case $option instanceof OptionGroup:
-					$results[] = [
-						'text'     => $option->get_label(),
-						'children' => $this->parse_options( $option->get_options() ),
-					];
-
-					break;
-				case $option instanceof Option:
-					$results[] = [
-						'id'   => $option->get_value(),
-						'text' => $option->get_label(),
-					];
-
-					break;
-			}
-		}
-
-		return $results;
-	}
-
 	public function __invoke() {
 		return [
-			'results'    => $this->parse_options( $this->options->get_copy() ),
+			'results'    => ArrayMapper::map( $this->options ),
 			'pagination' => [
 				'more' => $this->more,
 			],
