@@ -1,4 +1,4 @@
-<script type="ts">
+<script lang="ts">
     import {tick} from "svelte";
     import {fade} from 'svelte/transition';
 
@@ -74,7 +74,7 @@
 </script>
 
 <div class="acui-tooltip">
-	<div class="acui-tooltip-trigger" on:mouseenter={handleMouseEnter} on:mouseleave={handleMouseOut} bind:this={triggerEl}>
+	<div class="acui-tooltip-trigger" on:mouseenter={handleMouseEnter} on:mouseleave={handleMouseOut} bind:this={triggerEl} role="none">
 		<slot></slot>
 	</div>
 	{#if active }
