@@ -1,9 +1,10 @@
-<script type="ts">
+<script lang="ts">
     import {fade} from 'svelte/transition';
     import {onMount} from "svelte";
 
     export let trigger: HTMLElement;
     export let appendToBody: boolean = false;
+    export let maxHeight: string = null;
 
     let rootElement: HTMLElement;
     let menuElement: HTMLElement;
@@ -47,7 +48,7 @@
 </script>
 
 {#if appendToBody}
-	<div style="position: absolute; left:0; top:0;" bind:this={rootElement}>
+	<div style="position: absolute; left:0; top:0;" style:max-height={maxHeight} bind:this={rootElement}>
 
 	</div>
 {/if}
@@ -56,10 +57,11 @@
 		class:-append-to-body={appendToBody}
 		class:-bottom-left={!appendToBody && position ==='bottom-left'}
 		style={rootElementStyle}
+		style:max-height={maxHeight}
 		in:fade={{ duration: 100}} out:fade={{ duration: 100}}
 		bind:this={menuElement}
 >
-	<div class="acui-dropdown-content" role="list" aria-modal="true" on:itemSelect>
+	<div class="acui-dropdown-content" role="list" on:itemSelect>
 		<slot/>
 	</div>
 </div>
