@@ -1,4 +1,4 @@
-<script type="ts">
+<script lang="ts">
 
     export let value: string
     export let options: Array<{ value: string, label: string }>
