@@ -42,7 +42,7 @@ let config = {
 		extensions : ['.mjs', '.js', '.svelte', '.ts'],
 		mainFields : ['svelte', 'browser', 'module', 'main'],
 		alias : {
-			svelte : path.resolve( 'node_modules', 'svelte' ),
+			svelte: path.resolve('node_modules', 'svelte/src/runtime'),
 			ACUi : path.resolve( __dirname, 'ui' ),
 		},
 	},
